+++ src/payload/messagepack/tree_items.rs
@@ -12,7 +12,7 @@ pub fn tree_items(root: &Value) -> Vec<TreeItem<'_, JsonSelector>> {
     }
 }
 
-fn recurse(key: JsonSelector, value: &Value) -> TreeItem<JsonSelector> {
+fn recurse(key: JsonSelector, value: &Value) -> TreeItem<'_, JsonSelector> {
     match value {
         Value::Map(object) => {
             let text = key.to_string();