+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- CLI: Rename --client-key to --client-private-key. This highlights that the key is private more explicitly. A (hidden) alias for --client-key still exists so this is not breaking existing setups.
+
+### Fixed
+
+- Correctly print broker host:port with :
+
 ## [0.22.1] - 2025-04-08
 
 ### Added