+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.23.0] - 2026-05-17
+
 ### Added
 
 - Add --qos for selecting the MQTT Quality of Service for subscribing and publishing.